requirements.txt
setuptools
py-trello
